rename to use_late_for_private_fields_and_variables
