+++ example/all.yaml
@@ -166,7 +166,7 @@ linter:
     - use_function_type_syntax_for_parameters
     - use_is_even_rather_than_modulo
     - use_key_in_widget_constructors
-    - use_late
+    - use_late_for_private_fields_and_variables
     - use_raw_strings
     - use_rethrow_when_possible
     - use_setters_to_change_properties
