+++ test/rules/experiments/nnbd/rules/use_late_for_private_fields_and_variables.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-// test w/ `pub run test -N use_late`
+// test w/ `pub run test -N use_late_for_private_fields_and_variables`
 
 int? i; // OK
